- Flashlist implemtnation - Refactor code
