+++ package.json
@@ -20,6 +20,7 @@
     "@react-navigation/native": "^7.1.17",
     "@react-navigation/stack": "^7.4.8",
     "@reduxjs/toolkit": "2.9.0",
+    "@shopify/flash-list": "^2.0.3",
     "@testing-library/react-native": "^13.3.3",
     "axios": "1.11.0",
     "axios-logger": "^2.8.1",
