+++ src/components/movie-item/MovieItemComponent.tsx
@@ -1,18 +1,12 @@
 import React, { useState } from 'react';
-import {
-  FlatList,
-  FlatListProps,
-  Image,
-  ImageBackground,
-  TouchableOpacity,
-  View,
-} from 'react-native';
+import { Image, ImageBackground, TouchableOpacity, View } from 'react-native';
+import { FlashList, FlashListProps } from '@shopify/flash-list';
 import styles from './MovieItem.style.ts';
 import { AppConstants } from '../../constant/appConstants.ts';
 import { MovieItem } from '../../types/response/MovieItem.ts';
 
 interface MovieItemProps
-  extends Omit<FlatListProps<MovieItem>, 'data' | 'renderItem'> {
+  extends Omit<FlashListProps<MovieItem>, 'data' | 'renderItem'> {
   movies: Array<MovieItem>;
   onPress: (item: MovieItem) => void;
   loadMoreData: () => void;
@@ -55,7 +49,7 @@ const MovieItemComponent = ({
 
   return (
     <View style={styles.mainView}>
-      <FlatList
+      <FlashList
         style={styles.flatListContainer}
         data={movies}
         renderItem={movieItem}
