+++ src/screens/celebrity/popular/PopularCelebrity.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import styles from './PopularCelebrity.style.ts';
 import { View } from 'react-native';
-import { useLazyPopularCelebrityQuery } from '../../../redux/query/rtkQuery.ts';
+import { useLazyPopularCelebrityQuery } from '../../../service/rtk-query/rtkQuery.ts';
 import { NavigationProp, useNavigation } from '@react-navigation/native';
 import { RootStackParam } from '../../../types/navigation/NavigationTypes.ts';
 import { CelebrityItem } from '../../../types/response/Celebrity.ts';
